fix(api): update Docker run command and refactor import paths for agent tools
