+++ pkgs/api/src/lib/agent/tools/holesky/eigenlayerTool.ts
@@ -4,7 +4,7 @@ import "dotenv/config";
 import { http, createPublicClient, createWalletClient, parseUnits } from "viem";
 import { holesky } from "viem/chains";
 import { z } from "zod";
-import { createPrivyViemAccount } from "../../wallet/privy";
+import { createPrivyViemAccount } from "../../../wallet/privy";
 import { ERC20_ABI } from "../abis/erc20_abi";
 
 dotenv.config();
